refactor: 쇼핑 페이지 UI 피드백 반영
