+++ CHICCHIC/src/components/Counseling/CommentItem.tsx
@@ -9,7 +9,6 @@ type Props = {
   level?: 0 | 1; // 답글 / 댓글
   groupId?: number; // 댓글 목록
   replies?: Consult[]; // 답글 목록
-  profileImg: string;
   onSubmitReply?: (payload: {
     groupId: number;
     content: string;
@@ -20,7 +19,6 @@ export function CommentItem({
   item,
   level = 0,
   groupId,
-  profileImg,
   replies = [],
   onSubmitReply,
 }: Props) {
@@ -29,11 +27,15 @@ export function CommentItem({
 
   const avatarSize = isParent ? "w-12 h-12" : "w-10 h-10";
   const containerIndent = isParent ? "" : "ml-5";
-  const nameRowText = isParent ? "font-medium" : "text-sm text-gray-600";
+  const nameRowText = isParent ? "font-medium" : "text-md";
 
   return (
     <div className={`flex ${containerIndent}`}>
-      <img className={avatarSize} src={profileImg} alt="profile" />
+      <img
+        className={`rounded-full ${avatarSize}`}
+        src={item.profile}
+        alt="profile"
+      />
       <div className="flex flex-col w-full pl-6 sm:pl-10">
         {/* 작성자/시간 */}
         <div className={`flex items-center gap-2 ${nameRowText}`}>
